DO NOT MERGE: Print pool numbers
This shows that in the successful case, attempts 1 and 2 are on
different pools.

In the failing case, attempts 1 and 2 are on the same pool.

Seems like in the failing case we forget to "Returning object to pool".

Also, in the successful case, we eventually seem to return some object
to the previous pool.